Release/0.2.0

+++ sample/TypeScript/Classification/decode.ts
@@ -46,6 +46,11 @@ function main () {
     decodedData = Buffer.from(resultJson.Inferences[0].O, 'base64')
   } else {
     console.log('not inference result in this data')
+    if (fs.existsSync('./decoded_result_ObjectDetection.json')
+        && fs.lstatSync('./decoded_result_ObjectDetection.json').isSymbolicLink()) {
+      console.log('Can\'t open symbolic link file.')
+      return
+    }
     fs.writeFileSync('./decoded_result_Classification.json', JSON.stringify(resultJson, null, 4))
     console.log('write file : decoded_result_Classification.json')
     return
@@ -69,6 +74,11 @@ function main () {
     resultJson.Inferences[0][inferenceKey] = res
   }
 
+  if (fs.existsSync('./decoded_result_ObjectDetection.json')
+      && fs.lstatSync('./decoded_result_ObjectDetection.json').isSymbolicLink()) {
+    console.log('Can\'t open symbolic link file.')
+    return
+  }
   fs.writeFileSync('./decoded_result_Classification.json', JSON.stringify(resultJson, null, 4))
   console.log('write file : decoded_result_Classification.json')
 }
